# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: My API
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /users/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                email:
                  type: string
                password:
                  type: string
              required:
                - login
                - email
                - password
      responses:
        '200':
          description: User registered successfully
        '400':
          description: Registration error
  /users/login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
              required:
                - login
                - password
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid login or password
  /users/checkAuth:
    get:
      summary: Check if the user is authenticated
      responses:
        '200':
          description: Authenticated
        '401':
          description: Not authenticated
  /users/logout:
    post:
      summary: Logout the user
      responses:
        '200':
          description: Logged out successfully
  /words:
    get:
      summary: Get all words for the authenticated user
      responses:
        '200':
          description: List of words
        '401':
          description: Not authenticated
    post:
      summary: Add a new word
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                word:
                  type: string
                translation:
                  type: string
                next_review_time:
                  type: string
              required:
                - word
                - translation
      responses:
        '200':
          description: Word added successfully
        '401':
          description: Not authenticated
  /words/{wordId}:
    delete:
      summary: Delete a word by ID
      parameters:
        - in: path
          name: wordId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Word deleted successfully
        '401':
          description: Not authenticated
  /playlists:
    get:
      summary: Get all playlists for the authenticated user
      responses:
        '200':
          description: List of playlists
        '401':
          description: Not authenticated
    post:
      summary: Add a new playlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required:
                - name
      responses:
        '201':
          description: Playlist created successfully
        '401':
          description: Not authenticated
  /playlists/{playlistId}:
    get:
      summary: Get a playlist by ID
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Playlist data
        '404':
          description: Playlist not found
    delete:
      summary: Delete a playlist by ID
      parameters:
        - in: path
          name: playlistId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Playlist deleted successfully
        '401':
          description: Not authenticated